disable automatic relays, UI improvements
Viewer application will register as "prospective" clients,
with no relaying enabled. Relaying will only be available
for logged-in admin clients.

We improve the UI for the broadcaster, showing the view URL
and correctly updating the number of viewers. Viewers have
to be registered with the server, at least as prospective
viewers

we update the UI for log, showing more client details