UI enhancements and feedback form
We add the feedback form, together with the backend support -
we have the feedback model class, the XHR call endpoint, and
the admin displays.

We fix strings in the frontend for the app name, and we upgrade
the static libraries for jQuery, Bootstrap and angular bootstrap ui.

We change channel pages to derive from the home.html as to
preserve the top navigation forms.